+++ client/src/layouts/application/component.tsx
@@ -9,11 +9,11 @@ const ApplicationLayout: React.FC<ApplicationLayoutProps> = (props: ApplicationL
   const { children } = props;
 
   return (
-    <div className="h-screen flex bg-gray-50 overflow-hidden">
+    <div className="h-screen flex bg-gray-50">
       {/* Navigation */}
       <Sidebar />
 
-      <div className="flex-1 min-w-0 flex flex-col overflow-hidden">
+      <div className="flex-1 min-w-0 flex flex-col">
         {/* Mobile top navigation */}
         <HeaderMobile />